AC-1220: check if pod exist before forget it
and raise with message if not

+++ kuberdock-cli/kubecli/container/container.py
@@ -583,6 +583,8 @@ class KuberDock(KubeCtl):
         Deletes a given pending container
         """
         if self._data_path:
+            if not os.path.isfile(self._data_path):
+                raise SystemExit("Pod isn't found")
             os.unlink(self._data_path)
